pyproject.toml
[tool.poetry]
name = "policebot-source-code"
version = "0.1.0"
description = "A simple reporting bot I made for my Discord server."
authors = ["supremedoggo"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
